refactor(client): extract xstate model in a separate file to prevent circular dependencies
